+++ CHANGELOG.md
@@ -37,11 +37,14 @@ Semua perubahan yang dilakukan akan didokumentasikan di dalam file ini.
   - Menambah **edit booking** untuk mengedit pengajuan booking mereka.
   - Menambah **soft delete** agar data yang dihapus tidak hilang secara permanen, hanya tidak ditampilkan saja.
   - Fitur **booking detail** di detail model pada logika `assets/js/bookings/list.js`:
-    - Menampilkan informasi waktu update status (`StatusUpdatedAt`) dan waktu perubahan data terakhir (`UpdatedAt`).
+    - Menampilkan informasi waktu update status (`StatusUpdatedAt`), waktu perubahan data terakhir (`UpdatedAt`) dan alasan peminjaman (`Purpose`).
+  - Menambahkan **validasi** tidak dapat booking di *tanggal/jam masa lalu*.
+  - Menambahkan **validasi** agar tidak terjadi *bentrok jadwal* di ruangan yang sama.
+  - Menambah **filter** untuk melihat booking pada **tanggal tertentu atau ruangan tertentu**.
 - Fitur Admin Approval:
   - Mengisi halaman `admin-approval.html` yang berisikan daftar pengajuan yang memerlukan persetujuan.
   - Menambah `admin-approval.js` untuk otomatis memfilter booking dengan status **Pending**.
-  - Menggunakan endpoint `PATCH /api/bookings/{id}/status` untuk menyetujui dan menolak secara efisien, jadi tidak perlu mengirimkan kembali semua data booking ulang dengan endpotin `PUT`.
+  - Menggunakan endpoint `PATCH /api/bookings/{id}/status` untuk menyetujui dan menolak secara efisien, jadi tidak perlu mengirimkan kembali semua data booking ulang dengan endpoint `PUT`.
   - Menambahkan admin dapat mengubah status dari disetujui (`Approved`) menjadi ditolak (`Rejected`) dan sebaliknya.
 
   ## Fixed
@@ -49,3 +52,7 @@ Semua perubahan yang dilakukan akan didokumentasikan di dalam file ini.
   - **Dashboard**: Kini menampilkan data statistik *real-time* yang diambil dari API, menggantikan tampilan *dummy*.
   - Memperbaiki fungsi **tombol update status** yang ada di dashboard untuk admin (kesalahan penulisan fungsi).
   - Memperbaiki tampilan dashboard user.
+  -**Refactoring** Rooms:
+    - Mengubah `assets/js/rooms/user.js` menjadi `assets/js/rooms/list.js`, menghapus **detail button**, dan menyembunyikaan **buttons edit dan delete** (hanya admin yang bisa melakukan crud terhadap rooms).
+    - Mengubah `assets/js/rooms/user.js` menjadi `assets/js/rooms/form.js`.
+    - Memperbaiki fungsi filter/searching ruangan.